requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
openai
streamlit
requests_html
python-dotenv